+++ CHANGELOG.md
@@ -10,6 +10,11 @@
   - `sense_getter()`
   - `unescape_oewn_sense_key()`
   - `escape_oewn_sense_key()`
+* `wn.project.get_project()` ([#53])
+* `wn.project.Project` ([#53])
+* `wn.project.ResourceOnlyPackage` ([#53])
+* `path` property on `wn.project.Project` classes ([#53])
+* `delete` parameter on `wn.project.iterpackages()` ([#53])
 
 ## Changed
 
@@ -703,6 +708,7 @@ abandoned, but this is an entirely new codebase.
 [#40]: https://github.com/goodmami/wn/issues/40
 [#46]: https://github.com/goodmami/wn/issues/46
 [#47]: https://github.com/goodmami/wn/issues/47
+[#53]: https://github.com/goodmami/wn/issues/53
 [#55]: https://github.com/goodmami/wn/issues/55
 [#58]: https://github.com/goodmami/wn/issues/58
 [#59]: https://github.com/goodmami/wn/issues/59